+++ packages/atomic-react/src/components/search/components.ts
@@ -7,9 +7,13 @@ import {
   AtomicIcon as LitAtomicIcon,
   AtomicInsightInterface as LitAtomicInsightInterface,
   AtomicLayoutSection as LitAtomicLayoutSection,
+  AtomicLoadMoreResults as LitAtomicLoadMoreResults,
   AtomicPager as LitAtomicPager,
   AtomicQuerySummary as LitAtomicQuerySummary,
+  AtomicRecsInterface as LitAtomicRecsInterface,
   AtomicResult as LitAtomicResult,
+  AtomicResultChildrenTemplate as LitAtomicResultChildrenTemplate,
+  AtomicResultList as LitAtomicResultList,
   AtomicResultSectionActions as LitAtomicResultSectionActions,
   AtomicResultSectionBadges as LitAtomicResultSectionBadges,
   AtomicResultSectionBottomMetadata as LitAtomicResultSectionBottomMetadata,
@@ -20,11 +24,13 @@ import {
   AtomicResultSectionTitleMetadata as LitAtomicResultSectionTitleMetadata,
   AtomicResultSectionVisual as LitAtomicResultSectionVisual,
   AtomicResultsPerPage as LitAtomicResultsPerPage,
+  AtomicResultText as LitAtomicResultText,
   AtomicSearchBoxInstantResults as LitAtomicSearchBoxInstantResults,
   AtomicSearchBoxQuerySuggestions as LitAtomicSearchBoxQuerySuggestions,
   AtomicSearchBoxRecentQueries as LitAtomicSearchBoxRecentQueries,
   AtomicSearchInterface as LitAtomicSearchInterface,
   AtomicSearchLayout as LitAtomicSearchLayout,
+  AtomicText as LitAtomicText,
 } from '@coveo/atomic/components';
 import {createComponent} from '@lit/react';
 import React from 'react';
@@ -77,6 +83,12 @@ export const AtomicLayoutSection = createComponent({
   elementClass: LitAtomicLayoutSection,
 });
 
+export const AtomicLoadMoreResults = createComponent({
+  tagName: 'atomic-load-more-results',
+  react: React,
+  elementClass: LitAtomicLoadMoreResults,
+});
+
 export const AtomicPager = createComponent({
   tagName: 'atomic-pager',
   react: React,
@@ -89,10 +101,28 @@ export const AtomicQuerySummary = createComponent({
   elementClass: LitAtomicQuerySummary,
 });
 
+export const AtomicRecsInterface = createComponent({
+  tagName: 'atomic-recs-interface',
+  react: React,
+  elementClass: LitAtomicRecsInterface,
+});
+
 export const AtomicResult = createComponent({
   tagName: 'atomic-result',
   react: React,
   elementClass: LitAtomicResult,
+});
+
+export const AtomicResultChildrenTemplate = createComponent({
+  tagName: 'atomic-result-children-template',
+  react: React,
+  elementClass: LitAtomicResultChildrenTemplate,
+});
+
+export const AtomicResultList = createComponent({
+  tagName: 'atomic-result-list',
+  react: React,
+  elementClass: LitAtomicResultList,
 });
 
 export const AtomicResultSectionActions = createComponent({
@@ -155,6 +185,12 @@ export const AtomicResultsPerPage = createComponent({
   elementClass: LitAtomicResultsPerPage,
 });
 
+export const AtomicResultText = createComponent({
+  tagName: 'atomic-result-text',
+  react: React,
+  elementClass: LitAtomicResultText,
+});
+
 export const AtomicSearchBoxInstantResults = createComponent({
   tagName: 'atomic-search-box-instant-results',
   react: React,
@@ -183,4 +219,10 @@ export const AtomicSearchLayout = createComponent({
   tagName: 'atomic-search-layout',
   react: React,
   elementClass: LitAtomicSearchLayout,
+});
+
+export const AtomicText = createComponent({
+  tagName: 'atomic-text',
+  react: React,
+  elementClass: LitAtomicText,
 });