+++ packages/atomic/vitest-utils/testing-helpers/fixtures/atomic/commerce/atomic-commerce-recommendation-interface-fixture.ts
@@ -6,8 +6,8 @@ import {customElement, property, state} from 'lit/decorators.js';
 import {vi} from 'vitest';
 import type {CommerceStore} from '@/src/components/commerce/atomic-commerce-interface/store';
 import type {CommerceBindings} from '@/src/components/commerce/atomic-commerce-recommendation-interface/atomic-commerce-recommendation-interface.js';
+import {bindingsContext} from '@/src/components/common/context/bindings-context.js';
 import type {BaseAtomicInterface} from '@/src/components/common/interface/interface-common.js';
-import {bindingsContext} from '@/src/components/context/bindings-context.js';
 import {
   type InitializeEvent,
   markParentAsReady,